+++ src/main/routes/features/claimantResponse/claimantResponseTasklistController.ts
@@ -1,7 +1,6 @@
 import {Router} from 'express';
-import {CLAIM_DETAILS_URL, CLAIMANT_RESPONSE_TASK_LIST_URL} from '../../urls';
+import {CLAIMANT_RESPONSE_TASK_LIST_URL} from '../../urls';
 import {getDescription, getTitle} from '../../../services/features/response/taskListService';
-import {constructResponseUrlWithIdParams} from '../../../common/utils/urlFormatter';
 import {AppRequest} from 'models/AppRequest';
 import {getClaimById} from '../../../modules/utilityService';
 import {getClaimantResponseTaskLists} from 'services/features/claimantResponse/claimantResponseTasklistService/claimantResponseTasklistService';
@@ -12,16 +11,16 @@ const claimantResponseTasklistController = Router();
 claimantResponseTasklistController.get(CLAIMANT_RESPONSE_TASK_LIST_URL, async (req: AppRequest, res, next) => {
   try {
     const claimId = req.params.id;
+    req.session.claimId = claimId;
     const lang = req.query.lang ? req.query.lang : req.cookies.lang;
     const claim = await getClaimById(claimId, req);
     const taskLists = getClaimantResponseTaskLists(claim, claimId, lang);
-
-    req.session.claimId = claimId;
-
-    const title = getTitle(taskLists, lang);
-    const description = getDescription(taskLists, lang);
-    const claimDetailsUrl = constructResponseUrlWithIdParams(claimId, CLAIM_DETAILS_URL); // check this out if still required
-    res.render(claimantResponseTasklistViewPath, {taskLists, title, description, claim, claimDetailsUrl});
+    res.render(claimantResponseTasklistViewPath, {
+      claim,
+      taskLists,
+      title: getTitle(taskLists, lang),
+      description: getDescription(taskLists, lang),
+    });
   } catch (error) {
     next(error);
   }
