+++ src/main/services/features/claimantResponse/claimantResponseTasklistService/claimantResponseTasks/howDefendantRespondSectionTaks.ts
@@ -3,7 +3,7 @@ import {Claim} from 'common/models/claim';
 import {TaskStatus} from 'common/models/taskList/TaskStatus';
 import {isCaseDataMissing, isResponseTypeMissing} from 'common/utils/taskList/tasks/taskListHelpers';
 import {constructResponseUrlWithIdParams} from 'common/utils/urlFormatter';
-import {CLAIMANT_RESPONSE_REVIEW_DEFENDANTS_RESPONSE_URL, CLAIMANT_RESPONSE_SETTLE_ADMITTED_URL} from 'routes/urls';
+import {CLAIMANT_RESPONSE_REVIEW_DEFENDANTS_RESPONSE_URL} from 'routes/urls';
 import {Task} from 'models/taskList/task';
 
 export function getViewDefendantsReponseTask (claim: Claim, claimId: string, lang: string): Task {
@@ -18,17 +18,4 @@ export function getViewDefendantsReponseTask (claim: Claim, claimId: string, lan
     viewDefendantsReponseTask.status = TaskStatus.INCOMPLETE;
   }
   return viewDefendantsReponseTask;
-}
-
-export function getAcceptOrRejectDefendantAdmittedTask(claim: Claim, claimId: string, lang: string): Task {
-  // update complete status
-  const accceptOrRejectDefendantAdmittedTask  = {
-    description: t('CLAIMANT_RESPONSE_TASK_LIST.CHOOSE_WHAT_TODO_NEXT.ACCEPT_OR_REJECT_ADMITTED', {lng: lang}),
-    url: constructResponseUrlWithIdParams(claimId, CLAIMANT_RESPONSE_SETTLE_ADMITTED_URL),
-    status: TaskStatus.INCOMPLETE,
-  };
-  if (isCaseDataMissing(claim) || isResponseTypeMissing(claim?.respondent1)) {
-    accceptOrRejectDefendantAdmittedTask.status = TaskStatus.INCOMPLETE;
-  }
-  return accceptOrRejectDefendantAdmittedTask;
 }
